Move `ThrowingObserver` to the upper level and document
