+++ server/src/test/java/io/spine/server/transport/memory/SingleThreadInMemSubscriberTest.java
@@ -25,7 +25,7 @@ import io.spine.base.Identifier;
 import io.spine.grpc.MemoizingObserver;
 import io.spine.server.integration.ExternalMessage;
 import io.spine.server.transport.ChannelId;
-import io.spine.server.transport.memory.given.SingleThreadInMemSubscriberTestEnv.ThrowingObserver;
+import io.spine.server.transport.memory.given.ThrowingObserver;
 import io.spine.testing.logging.MuteLogging;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
